Add new state 'restarted' to make the logic for restarting servers more straightforward.
Unique names: One can now query the API of an already existing server with 'servername'. This will return an error if there is more than one server that has this name (non-case-sensitive). If your servers all have the same name, then either change the server names in Cloud or use the instanceid method.